requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary
Flask-Migrate==4.0.5
Werkzeug==2.0.3
python-dotenv
SQLAlchemy==1.4.36
